[ZODC-XXX][1.1.22] REMOVE DUPLICATE CALCULATE REQUETS PROGRESS IN APPROVE REJECT NODE

+++ internal/app/dto/responses/request_response.go
@@ -25,7 +25,7 @@ type RequestResponse struct {
 	Title        string                `json:"title"`
 	Status       string                `json:"status"`
 	ParentKey    int32                 `json:"parentKey"`
-	Progress     int32                 `json:"progress"`
+	Progress     float32               `json:"progress"`
 	CurrentTasks []CurrentTaskResponse `json:"currentTasks,omitempty"`
 	SprintId     int32                 `json:"sprintId"`
 	StartedAt    *time.Time            `json:"startedAt"`
